+++ extensions/dear-imnodes/src/context/editor/build.rs
@@ -3,9 +3,9 @@ use crate::sys;
 
 impl<'ui> NodeEditor<'ui> {
     /// Begin a node
-    pub fn node(&self, id: i32) -> NodeToken<'_> {
+    pub fn node(&self, id: crate::NodeId) -> NodeToken<'_> {
         self.bind();
-        unsafe { sys::imnodes_BeginNode(id) };
+        unsafe { sys::imnodes_BeginNode(id.raw()) };
         NodeToken {
             scope: self.scope(),
             _phantom: std::marker::PhantomData,
@@ -13,9 +13,9 @@ impl<'ui> NodeEditor<'ui> {
     }
 
     /// Begin an input attribute pin
-    pub fn input_attr(&self, id: i32, shape: crate::PinShape) -> AttributeToken<'_> {
+    pub fn input_attr(&self, id: crate::PinId, shape: crate::PinShape) -> AttributeToken<'_> {
         self.bind();
-        unsafe { sys::imnodes_BeginInputAttribute(id, shape as sys::ImNodesPinShape) };
+        unsafe { sys::imnodes_BeginInputAttribute(id.raw(), shape as sys::ImNodesPinShape) };
         AttributeToken {
             kind: AttrKind::Input,
             scope: self.scope(),
@@ -24,9 +24,9 @@ impl<'ui> NodeEditor<'ui> {
     }
 
     /// Begin an output attribute pin
-    pub fn output_attr(&self, id: i32, shape: crate::PinShape) -> AttributeToken<'_> {
+    pub fn output_attr(&self, id: crate::PinId, shape: crate::PinShape) -> AttributeToken<'_> {
         self.bind();
-        unsafe { sys::imnodes_BeginOutputAttribute(id, shape as sys::ImNodesPinShape) };
+        unsafe { sys::imnodes_BeginOutputAttribute(id.raw(), shape as sys::ImNodesPinShape) };
         AttributeToken {
             kind: AttrKind::Output,
             scope: self.scope(),
@@ -35,9 +35,9 @@ impl<'ui> NodeEditor<'ui> {
     }
 
     /// Begin a static attribute region
-    pub fn static_attr(&self, id: i32) -> AttributeToken<'_> {
+    pub fn static_attr(&self, id: crate::PinId) -> AttributeToken<'_> {
         self.bind();
-        unsafe { sys::imnodes_BeginStaticAttribute(id) };
+        unsafe { sys::imnodes_BeginStaticAttribute(id.raw()) };
         AttributeToken {
             kind: AttrKind::Static,
             scope: self.scope(),
@@ -46,8 +46,8 @@ impl<'ui> NodeEditor<'ui> {
     }
 
     /// Draw a link between two attributes
-    pub fn link(&self, id: i32, start_attr: i32, end_attr: i32) {
+    pub fn link(&self, id: crate::LinkId, start_attr: crate::PinId, end_attr: crate::PinId) {
         self.bind();
-        unsafe { sys::imnodes_Link(id, start_attr, end_attr) }
+        unsafe { sys::imnodes_Link(id.raw(), start_attr.raw(), end_attr.raw()) }
     }
 }